refactor pack packaging
