+++ src/main/java/harmonised/pmmo/setup/CommonSetup.java
@@ -7,7 +7,6 @@ import harmonised.pmmo.compat.ftb_quests.FTBQHandler;
 import harmonised.pmmo.config.Config;
 import harmonised.pmmo.config.readers.CoreLoader;
 import harmonised.pmmo.config.readers.ExecutableListener;
-import harmonised.pmmo.config.scripting.Scripting;
 import harmonised.pmmo.core.Core;
 import harmonised.pmmo.core.perks.PerkRegistration;
 import harmonised.pmmo.features.loot_modifiers.SkillUpTrigger;
@@ -18,17 +17,17 @@ import harmonised.pmmo.setup.datagen.EntityTagProvider;
 import harmonised.pmmo.setup.datagen.ItemTagProvider;
 import harmonised.pmmo.setup.datagen.LangProvider;
 import harmonised.pmmo.setup.datagen.LangProvider.Locale;
-import harmonised.pmmo.setup.datagen.defaultpacks.DefaultBiomeConfigProvider;
-import harmonised.pmmo.setup.datagen.defaultpacks.DefaultBlockConfigProvider;
-import harmonised.pmmo.setup.datagen.defaultpacks.DefaultDimConfigProvider;
-import harmonised.pmmo.setup.datagen.defaultpacks.DefaultEntityConfigProvider;
-import harmonised.pmmo.setup.datagen.defaultpacks.DefaultGLMProvider;
-import harmonised.pmmo.setup.datagen.defaultpacks.DefaultItemConfigProvider;
-import harmonised.pmmo.setup.datagen.defaultpacks.EasyBlockConfigProvider;
-import harmonised.pmmo.setup.datagen.defaultpacks.EasyConfigProvider;
-import harmonised.pmmo.setup.datagen.defaultpacks.EasyGLMProvider;
-import harmonised.pmmo.setup.datagen.defaultpacks.EasyItemConfigProvider;
-import harmonised.pmmo.setup.datagen.defaultpacks.HardcoreGLMProvider;
+import harmonised.pmmo.setup.datagen.defaultpack.DefaultBiomeConfigProvider;
+import harmonised.pmmo.setup.datagen.defaultpack.DefaultBlockConfigProvider;
+import harmonised.pmmo.setup.datagen.defaultpack.DefaultDimConfigProvider;
+import harmonised.pmmo.setup.datagen.defaultpack.DefaultEntityConfigProvider;
+import harmonised.pmmo.setup.datagen.defaultpack.DefaultGLMProvider;
+import harmonised.pmmo.setup.datagen.defaultpack.DefaultItemConfigProvider;
+import harmonised.pmmo.setup.datagen.easypack.EasyBlockConfigProvider;
+import harmonised.pmmo.setup.datagen.easypack.EasyConfigProvider;
+import harmonised.pmmo.setup.datagen.easypack.EasyGLMProvider;
+import harmonised.pmmo.setup.datagen.easypack.EasyItemConfigProvider;
+import harmonised.pmmo.setup.datagen.hardcorepack.HardcoreGLMProvider;
 import harmonised.pmmo.util.MsLoggy;
 import harmonised.pmmo.util.MsLoggy.LOG_CODE;
 import harmonised.pmmo.util.Reference;
@@ -46,7 +45,6 @@ import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.LogicalSide;
 import net.neoforged.fml.ModList;
 import net.neoforged.fml.common.EventBusSubscriber;
-import net.neoforged.fml.common.Mod;
 import net.neoforged.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.neoforged.neoforge.data.event.GatherDataEvent;
 import net.neoforged.neoforge.event.AddReloadListenerEvent;
