+++ src/main/java/harmonised/pmmo/setup/datagen/defaultpack/DefaultGLMProvider.java
@@ -1,4 +1,4 @@
-package harmonised.pmmo.setup.datagen.defaultpacks;
+package harmonised.pmmo.setup.datagen.defaultpack;
 
 import harmonised.pmmo.setup.datagen.GLMProvider;
 import net.minecraft.core.HolderLookup;
@@ -7,9 +7,9 @@ import net.minecraft.data.PackOutput;
 import java.nio.file.Path;
 import java.util.concurrent.CompletableFuture;
 
-public class HardcoreGLMProvider extends GLMProvider {
-    private static final Path path = Path.of("resourcepacks/hardcore/data");
-    public HardcoreGLMProvider(PackOutput gen, CompletableFuture<HolderLookup.Provider> registries) {
+public class DefaultGLMProvider extends GLMProvider {
+    private static final Path path = Path.of("resourcepacks/default/data");
+    public DefaultGLMProvider(PackOutput gen, CompletableFuture<HolderLookup.Provider> registries) {
         super(gen, path, registries);
     }
